Development

+++ src/app_timer.c
@@ -17,7 +17,9 @@ bool app_timer_exceed(uint32_t ref, uint32_t ms) {
         tt = (tt - app_timer.timerAPPEvt->period) * -1;
     } else tt = ref - t;
 
-	return (tt > ms);
+//    printf("tt: %d, t: %d, ref: %d, ms: %d, (tt > ms): %d\r\n", tt, t, ref, ms, ((uint32_t)tt > ms));
+
+	return ((uint32_t)tt > ms);
 }
 
 uint32_t app_timeout_get() {
